Add sorting and filtering to job queries
This commit has two goals:
- Filter jobs I was rejected from (for tidiness and morale) in the
  sidebar of the Jobs tab
- Sort Open Jobs in the Company tab to make it easier to see what I'm
  waiting on and what to follow up on

The main question is whether to execute these changes in the DB or in
the application. For now, I'm going to try to keep the DB simple and
straightforward. DB is for data, application is for everything else.

To that end, both the sorting and filtering changes were made to app
code.

Upsides:
- I don't have DB versioning and update code yet, so it's a small pain
  to make schema changes.
- Complexity stays in the app and keeps the DB nice and dumb (no weird
  multiple ORDER BY clauses or a new table for sorting or a new int
column that duplicates the job_status)

Downsides:
- Adds more conditionals to the app code. I was hoping soon to refactor
  a bunch of the code and have fewer conditional branches based on
tab/table...but I suppose I could abstract the ordering too.
- It isn't the most efficient in that one needs to go through the list
  more than once. That said, these lists shouldn't be more than a
handful of jobs anyway, so bad loops aren't the worst and I'm sure there
are better approaches.
- Separates some of the data away from the database. Job status order IS
  part of the information, and part of me thinks it would be better to
keep that info in the database. On the other hand, I may want to change
the order later (like perhaps I want to see jobs I have interviews for
at the top of the list.